Fixed two programs to run in VS

+++ prog2.c
@@ -60,7 +60,7 @@ void drawCircle(int xc, int yc, int r)
 
 int getRadius(int x1, int y1, int x2, int y2)
 {
-    return (int)sqrt(((x2-x1)*(x2-x1))+((y2-y1)*(y2-y1)));
+    return (int)sqrt(float((x2-x1)*(x2-x1))+((y2-y1)*(y2-y1)));
 }
 
 void display() {}
